first prototype complete
changes:
-file renamed to flightbooking.cpp
-flight.h and flight.cpp added with flight and plane classes
-addBooking will generate a seating chart if none exists
-functionality to choose flight and seat number and have it booked added
-fixed seatID algorithm

limitations:
-file must exist to begin with otherwise the shown seating chart will
not be consistent with actual file contents
-no flight IDs are shown so user must already know this information

future improvements:
-detect if seat is already booked (will be easiest if this is a
function)
-create seperate folder for seating charts
-add referencing system for the generated seating chart so users know
which co-ordinates correspond to each seat
