+++ BluebrookAirport/BluebrookAirport/BluebrookAirport.vcxproj
@@ -145,11 +145,13 @@
     <Text Include="ReadMe.txt" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="flight.h" />
     <ClInclude Include="stdafx.h" />
     <ClInclude Include="targetver.h" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="Plane+Flight_Class.cpp" />
+    <ClCompile Include="flight.cpp" />
+    <ClCompile Include="flightbooking.cpp" />
     <ClCompile Include="PlaneNumberGenerator.cpp" />
     <ClCompile Include="stdafx.cpp">
       <PrecompiledHeader Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">Create</PrecompiledHeader>
