+++ BluebrookAirport/BluebrookAirport/BluebrookAirport.vcxproj.filters
@@ -24,6 +24,9 @@
     <ClInclude Include="targetver.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="flight.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="stdafx.cpp">
@@ -32,7 +35,10 @@
     <ClCompile Include="PlaneNumberGenerator.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="Plane+Flight_Class.cpp">
+    <ClCompile Include="flight.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
+    <ClCompile Include="flightbooking.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
