--- BluebrookAirport/BluebrookAirport/FFF999.txt
@@ -1,25 +1,0 @@
-O
-O
-O
-O
-O
-O
-O
-O
-X
-O
-X
-O
-X
-O
-O
-O
-O
-O
-O
-O
-X
-O
-O
-O
-O
