+++ BluebrookAirport/BluebrookAirport/x64/Debug/BluebrookAirport.log
@@ -1,3 +1,3 @@
-﻿  Plane+Flight_Class.cpp
+﻿  flightbooking.cpp
   BluebrookAirport.vcxproj -> C:\Users\slcra\Desktop\Software Coursework\Software-Design-Imp\BluebrookAirport\x64\Debug\BluebrookAirport.exe
   BluebrookAirport.vcxproj -> C:\Users\slcra\Desktop\Software Coursework\Software-Design-Imp\BluebrookAirport\x64\Debug\BluebrookAirport.pdb (Full PDB)
